Split widget: sparkline
