+++ extension/devtools/app/lib/main.dart
@@ -32,6 +32,7 @@ part 'widgets/health_card.dart';
 part 'widgets/health_bar.dart';
 part 'widgets/chart_placeholder.dart';
 part 'widgets/mini_chart.dart';
+part 'widgets/sparkline.dart';
 part 'widgets/metric_tile.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
@@ -3209,35 +3210,6 @@ class _PanelInfo {
   final String title;
   final String description;
   final List<String> bullets;
-}
-
-class _Sparkline extends StatelessWidget {
-  const _Sparkline({required this.color});
-
-  final Color color;
-
-  @override
-  Widget build(BuildContext context) {
-    const bars = [0.3, 0.5, 0.7, 0.4, 0.6, 0.8, 0.5, 0.9];
-    return SizedBox(
-      height: 28,
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        children: [
-          for (final value in bars)
-            Container(
-              width: 10,
-              height: 28 * value,
-              margin: const EdgeInsets.only(right: 4),
-              decoration: BoxDecoration(
-                color: color.withValues(alpha: 0.4),
-                borderRadius: BorderRadius.circular(6),
-              ),
-            ),
-        ],
-      ),
-    );
-  }
 }
 
 class _TimelineRow extends StatelessWidget {
